refactor: extract duplicated inDarkMode() helper to ThemeUtils
The modernization campaign re-introduced an identical private
isInDarkMode(context) in both GoogleComposeAdapter and
MapLibreComposeAdapter. Hoist the single copy into
ThemeUtils.isInDarkMode(context) and call it from both adapters,
dropping the now-unused AppCompatDelegate/Configuration imports.

Rebased onto current main from the original PR #1545, whose target
files (BaseMapFragment/MapLibreMapFragment/UIUtils) no longer exist.
